Fix table stats delayed appearance
Rename MainScreen to Screener and move it to a separate file

Fix some typos
